pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "freewrite"
version = "0.1.0"
description = "FreeWrite Backend API"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.2",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.3.0",
    "pydantic-settings>=2.0.3",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "greenlet>=3.2.1",
    "pytz>=2023.3",
    "psycopg2-binary>=2.9.10",
    "openai>=1.76.2",
    "redis>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 
